// _src/gaf-master-dashboard/types.ts
export interface SubTheme {
  name: string;
  count: number;
  jobs: { name: string; count: number }[];
}

export interface ProcessedTheme {
  name: string;
  count: number;
  percentage: string;
  subThemes: SubTheme[];
}